+++ apps/ledger-live-mobile/src/actions/types.ts
@@ -24,6 +24,7 @@ import type {
   State,
   WalletConnectState,
   SwapStateType,
+  DynamicContentState,
 } from "../reducers/types";
 import type { Unpacked } from "../types/helpers";
 
@@ -162,6 +163,27 @@ export type NotificationsPayload =
   | NotificationsSetEventTriggeredPayload
   | NotificationsSetDataOfUserPayload;
 
+// === DYNAMIC CONTENT ACTIONS ===
+
+export enum DynamicContentActionTypes {
+  DYNAMIC_CONTENT_SET_WALLET_CARDS = "DYNAMIC_CONTENT_SET_WALLET_CARDS",
+  DYNAMIC_CONTENT_SET_ASSET_CARDS = "DYNAMIC_CONTENT_SET_ASSET_CARDS",
+}
+
+export type DynamicContentSetWalletCardsPayload = Pick<
+  DynamicContentState,
+  "walletCards"
+>;
+
+export type DynamicContentSetAssetCardsPayload = Pick<
+  DynamicContentState,
+  "assetsCards"
+>;
+
+export type DynamicContentPayload =
+  | DynamicContentSetWalletCardsPayload
+  | DynamicContentSetAssetCardsPayload;
+
 // === RATINGS ACTIONS ===
 
 export enum RatingsActionTypes {
@@ -209,6 +231,7 @@ export enum SettingsActionTypes {
   SETTINGS_SET_PAIRS = "SETTINGS_SET_PAIRS",
   SETTINGS_SET_SELECTED_TIME_RANGE = "SETTINGS_SET_SELECTED_TIME_RANGE",
   SETTINGS_COMPLETE_ONBOARDING = "SETTINGS_COMPLETE_ONBOARDING",
+  SETTINGS_COMPLETE_CUSTOM_IMAGE_FLOW = "SETTINGS_COMPLETE_CUSTOM_IMAGE_FLOW",
   SETTINGS_INSTALL_APP_FIRST_TIME = "SETTINGS_INSTALL_APP_FIRST_TIME",
   SETTINGS_SET_READONLY_MODE = "SETTINGS_SET_READONLY_MODE",
   SETTINGS_SET_EXPERIMENTAL_USB_SUPPORT = "SETTINGS_SET_EXPERIMENTAL_USB_SUPPORT",
@@ -224,6 +247,7 @@ export enum SettingsActionTypes {
   SETTINGS_SET_THEME = "SETTINGS_SET_THEME",
   SETTINGS_SET_OS_THEME = "SETTINGS_SET_OS_THEME",
   SETTINGS_SET_CAROUSEL_VISIBILITY = "SETTINGS_SET_CAROUSEL_VISIBILITY",
+  SETTINGS_SET_DISMISSED_DYNAMIC_CARDS = "SETTINGS_SET_DISMISSED_DYNAMIC_CARDS",
   SETTINGS_SET_DISCREET_MODE = "SETTINGS_SET_DISCREET_MODE",
   SETTINGS_SET_LANGUAGE = "SETTINGS_SET_LANGUAGE",
   SETTINGS_SET_LOCALE = "SETTINGS_SET_LOCALE",
@@ -235,6 +259,7 @@ export enum SettingsActionTypes {
   ADD_STARRED_MARKET_COINS = "ADD_STARRED_MARKET_COINS",
   REMOVE_STARRED_MARKET_COINS = "REMOVE_STARRED_MARKET_COINS",
   SET_LAST_CONNECTED_DEVICE = "SET_LAST_CONNECTED_DEVICE",
+  SET_CUSTOM_IMAGE_BACKUP = "SET_CUSTOM_IMAGE_BACKUP",
   SET_HAS_ORDERED_NANO = "SET_HAS_ORDERED_NANO",
   SET_MARKET_REQUEST_PARAMS = "SET_MARKET_REQUEST_PARAMS",
   SET_MARKET_COUNTER_CURRENCY = "SET_MARKET_COUNTER_CURRENCY",
@@ -243,6 +268,7 @@ export enum SettingsActionTypes {
   SET_FIRST_CONNECTION_HAS_DEVICE = "SET_FIRST_CONNECTION_HAS_DEVICE",
   SET_NOTIFICATIONS = "SET_NOTIFICATIONS",
   RESET_SWAP_LOGIN_AND_KYC_DATA = "RESET_SWAP_LOGIN_AND_KYC_DATA",
+  WALLET_TAB_NAVIGATOR_LAST_VISITED_TAB = "WALLET_TAB_NAVIGATOR_LAST_VISITED_TAB",
 }
 
 export type SettingsImportPayload = Partial<SettingsState>;
@@ -313,6 +339,10 @@ export type SettingsSetCarouselVisibilityPayload = Pick<
   SettingsState,
   "carouselVisibility"
 >;
+export type SettingsSetDismissedDynamicCardsPayload = Pick<
+  SettingsState,
+  "dismissedDynamicCards"
+>;
 export type SettingsSetDiscreetModePayload = Pick<
   SettingsState,
   "discreetMode"
@@ -346,6 +376,10 @@ export type SettingsRemoveStarredMarketcoinsPayload = {
 export type SettingsSetLastConnectedDevicePayload = {
   lastConnectedDevice: Device;
 };
+export type SettingsSetCustomImageBackupPayload = {
+  hex: string;
+  hash: string;
+};
 export type SettingsSetHasOrderedNanoPayload = Pick<
   SettingsState,
   "hasOrderedNano"
@@ -376,6 +410,10 @@ export type SettingsSetFirstConnectHasDeviceUpdatedPayload = Pick<
 export type SettingsSetNotificationsPayload = {
   notifications: Partial<SettingsState["notifications"]>;
 };
+export type SettingsSetWalletTabNavigatorLastVisitedTabPayload = Pick<
+  SettingsState,
+  "walletTabNavigatorLastVisitedTab"
+>;
 export type SettingsDangerouslyOverrideStatePayload = State;
 export type SettingsPayload =
   | SettingsImportPayload