+++ apps/ledger-live-mobile/src/components/RootNavigator/types/SettingsNavigator.ts
@@ -51,4 +51,5 @@ export type SettingsNavigatorStackParamList = {
   [ScreenName.PostOnboardingDebugScreen]: undefined;
   [ScreenName.DebugMultiAppInstall]: undefined;
   [ScreenName.DebugStoryly]: undefined;
+  [ScreenName.DebugFetchCustomImage]: undefined;
 };