+++ apps/ledger-live-mobile/src/components/RootNavigator/types/SwapNavigator.ts
@@ -66,7 +66,7 @@ export type SwapNavigatorParamList = {
   [ScreenName.SwapSelectProvider]: {
     provider?: string;
     swap: SwapDataType;
-    selectedId?: string;
+    selectedRate: ExchangeRate | undefined;
   };
   [ScreenName.SwapSelectFees]: {
     accountId?: string;