+++ apps/ledger-live-mobile/src/reducers/types.ts
@@ -17,6 +17,8 @@ import {
 import { Transaction } from "@ledgerhq/live-common/generated/types";
 import type { EventTrigger, DataOfUser } from "../logic/notifications";
 import type { RatingsHappyMoment, RatingsDataOfUser } from "../logic/ratings";
+import { WalletTabNavigatorStackParamList } from "../components/RootNavigator/types/WalletTabNavigator";
+import { WalletContentCard, AssetContentCard } from "../dynamicContent/types";
 
 // === ACCOUNT STATE ===
 
@@ -69,7 +71,7 @@ export type DeviceLike = {
   name: string;
   deviceInfo?: DeviceInfo;
   appsInstalled?: number;
-  modelId?: DeviceModelId;
+  modelId: DeviceModelId;
 };
 
 export type BleState = {
@@ -94,6 +96,15 @@ export type NotificationsState = {
    * If true, it means another modal is already opened or being opened
    */
   isPushNotificationsModalLocked: boolean;
+};
+
+// === DYNAMIC CONTENT STATE ===
+
+export type DynamicContentState = {
+  /** Dynamic content cards displayed in the Wallet Page */
+  walletCards: WalletContentCard[];
+  /** Dynamic content cards displayed in an Asset Page */
+  assetsCards: AssetContentCard[];
 };
 
 // === RATINGS STATE ===
@@ -153,6 +164,7 @@ export type SettingsState = {
   pairExchanges: Record<string, string | null | undefined>;
   selectedTimeRange: PortfolioRange;
   orderAccounts: string;
+  hasCompletedCustomImageFlow: boolean;
   hasCompletedOnboarding: boolean;
   hasInstalledAnyApp: boolean;
   readOnlyModeEnabled: boolean;
@@ -168,6 +180,7 @@ export type SettingsState = {
   theme: Theme;
   osTheme: string | null | undefined;
   carouselVisibility: number | Record<string, boolean>;
+  dismissedDynamicCards: string[];
   // number is the legacy type from LLM V2
   discreetMode: boolean;
   language: string;
@@ -190,13 +203,15 @@ export type SettingsState = {
   sensitiveAnalytics: boolean;
   firstConnectionHasDevice: boolean | null;
   firstConnectHasDeviceUpdated: boolean | null;
+  customImageBackup?: { hex: string; hash: string };
   notifications: NotificationsSettings;
+  walletTabNavigatorLastVisitedTab: keyof WalletTabNavigatorStackParamList;
 };
 
 export type NotificationsSettings = {
-  allowed: boolean;
-  announcements: boolean;
-  recommendations: boolean;
+  areNotificationsAllowed: boolean;
+  announcementsCategory: boolean;
+  recommendationsCategory: boolean;
 };
 
 // === WALLET CONNECT STATE ===
@@ -223,6 +238,7 @@ export type State = {
   appstate: AppState;
   ble: BleState;
   ratings: RatingsState;
+  dynamicContent: DynamicContentState;
   notifications: NotificationsState;
   swap: SwapStateType;
   walletconnect: WalletConnectState;