+++ apps/ledger-live-mobile/src/screens/Settings/Notifications/index.tsx
@@ -20,6 +20,12 @@ import type { State } from "../../../reducers/types";
 import useNotifications from "../../../logic/notifications";
 import { updateUserPreferences } from "../../../notifications/braze";
 
+const notificationsMapping = {
+  areNotificationsAllowed: "allowed",
+  announcementsCategory: "announcements",
+  recommendationsCategory: "recommendations",
+};
+
 type NotificationRowProps = {
   disabled?: boolean;
   notificationKey: keyof State["settings"]["notifications"];
@@ -36,7 +42,8 @@ function NotificationSettingsRow({
 
   const { t } = useTranslation();
 
-  const capitalizedKey = capitalize(notificationKey);
+  const translationKey = notificationsMapping[notificationKey];
+  const capitalizedKey = capitalize(translationKey);
 
   const onChange = useCallback(
     (value: boolean) => {
@@ -58,8 +65,8 @@ function NotificationSettingsRow({
   return (
     <SettingsRow
       event={`${capitalizedKey}Row`}
-      title={t(`settings.notifications.${notificationKey}.title`)}
-      desc={t(`settings.notifications.${notificationKey}.desc`)}
+      title={t(`settings.notifications.${translationKey}.title`)}
+      desc={t(`settings.notifications.${translationKey}.desc`)}
       label={label}
     >
       <Switch
@@ -112,7 +119,7 @@ function NotificationsSettings() {
     updateUserPreferences(notifications);
   }, [notifications]);
 
-  const disableSubSettings = !notifications.allowed;
+  const disableSubSettings = !notifications.areNotificationsAllowed;
 
   const platformData = useMemo(
     () =>
@@ -168,27 +175,33 @@ function NotificationsSettings() {
           ) : null}
           <Box opacity={isNotifPermissionEnabled ? 1 : 0.2}>
             <NotificationSettingsRow
-              notificationKey={"allowed"}
+              notificationKey={"areNotificationsAllowed"}
               disabled={!isNotifPermissionEnabled}
             />
           </Box>
           <Box
             opacity={
-              isNotifPermissionEnabled && notifications.allowed ? 1 : 0.2
+              isNotifPermissionEnabled && notifications.areNotificationsAllowed
+                ? 1
+                : 0.2
             }
           >
             <NotificationSettingsRow
-              notificationKey={"announcements"}
+              notificationKey={"announcementsCategory"}
               disabled={disableSubSettings}
             />
             <NotificationSettingsRow
-              notificationKey={"recommendations"}
+              notificationKey={"recommendationsCategory"}
               disabled={disableSubSettings}
             />
           </Box>
           <Box m={6}>
             <Text
-              color={notifications.allowed ? "neutral.c40" : "neutral.c70"}
+              color={
+                notifications.areNotificationsAllowed
+                  ? "neutral.c40"
+                  : "neutral.c70"
+              }
               variant={"bodyLineHeight"}
               textAlign="center"
             >