+++ libs/ledgerjs/packages/errors/src/index.ts
@@ -52,6 +52,7 @@ export const DeviceSocketFail = createCustomErrorClass("DeviceSocketFail");
 export const DeviceSocketNoBulkStatus = createCustomErrorClass(
   "DeviceSocketNoBulkStatus"
 );
+export const LockedDeviceError = createCustomErrorClass("LockedDeviceError");
 export const DisconnectedDevice = createCustomErrorClass("DisconnectedDevice");
 export const DisconnectedDeviceDuringOperation = createCustomErrorClass(
   "DisconnectedDeviceDuringOperation"
@@ -236,6 +237,36 @@ export const DBWrongPassword = createCustomErrorClass("DBWrongPassword");
 export const DBNotReset = createCustomErrorClass("DBNotReset");
 
 /**
+ * Type of a Transport error used to represent all equivalent errors coming from all possible implementation of Transport
+ */
+export enum HwTransportErrorType {
+  Unknown = 0,
+  BleLocationServicesDisabled = 1,
+  BleBluetoothUnauthorized = 2,
+  BleScanStartFailed = 3,
+}
+
+/**
+ * Represents an error coming from any Transport implementation.
+ *
+ * Needed to map a specific implementation error into an error that
+ * can be managed by any code unaware of the specific Transport implementation
+ * that was used.
+ */
+export class HwTransportError extends Error {
+  type: HwTransportErrorType;
+
+  constructor(type: HwTransportErrorType, message: string) {
+    super(message);
+    this.name = "HwTransportError";
+    this.type = type;
+
+    // Needed as long as we target < ES6
+    Object.setPrototypeOf(this, HwTransportError.prototype);
+  }
+}
+
+/**
  * TransportError is used for any generic transport errors.
  * e.g. Error thrown when data received by exchanges are incorrect or if exchanged failed to communicate with the device for various reason.
  */
@@ -289,6 +320,8 @@ export const StatusCodes = {
   LICENSING: 0x6f42,
   HALTED: 0x6faa,
   LOCKED_DEVICE: 0x5515,
+  CUSTOM_IMAGE_EMPTY: 0x662e,
+  CUSTOM_IMAGE_BOOTLOADER: 0x662f,
 };
 
 export function getAltStatusMessage(code: number): string | undefined | null {
@@ -319,13 +352,20 @@ export function getAltStatusMessage(code: number): string | undefined | null {
  * the error.statusCode is one of the `StatusCodes` exported by this library.
  */
 export function TransportStatusError(statusCode: number): void {
-  this.name = "TransportStatusError";
   const statusText =
     Object.keys(StatusCodes).find((k) => StatusCodes[k] === statusCode) ||
     "UNKNOWN_ERROR";
   const smsg = getAltStatusMessage(statusCode) || statusText;
   const statusCodeStr = statusCode.toString(16);
-  this.message = `Ledger device: ${smsg} (0x${statusCodeStr})`;
+  const message = `Ledger device: ${smsg} (0x${statusCodeStr})`;
+
+  // Maps to a LockedDeviceError
+  if (statusCode === StatusCodes.LOCKED_DEVICE) {
+    throw new LockedDeviceError(message);
+  }
+
+  this.name = "TransportStatusError";
+  this.message = message;
   this.stack = new Error().stack;
   this.statusCode = statusCode;
   this.statusText = statusText;